add more explanations

+++ diff.py
@@ -127,7 +127,15 @@ def _make_js_list():
 	print('Job Completed, js_result.txt has been created')
 
 	print('Cleaning up previous js_files')
-	shutil.rmtree('./js_files')
+
+	if os.path.exists(os.path.dirname('./js_files')):
+	    shutil.rmtree('./js_files')
+	    try:
+		os.makedirs(os.path.dirname('./js_files'))
+	    except OSError as exc: # Guard against race condition
+		if exc.errno != errno.EEXIST:
+		    raise
+	
 	for index, jscript in enumerate(js_list):
 		try:
 			r = requests.get(jscript)
@@ -154,6 +162,7 @@ def _recursive_diff(folder1, folder2):
 	f = open('result.txt','w')
 	if len(html_files_folder1) == len(html_files_folder2):
 		for x in range(0,len(html_files_folder1)-1):
+			print("Comparing "+ str(os.path.basename(html_files_folder1[x])) +"(file #1) and "+ str(os.path.basename(html_files_folder2[x]))+"(file #2) to see if any injected things inside the file #2.")
 			f.write("\r\r\r\r")
 			f.write("---[+][+][+]----Injected scripts into the-------"+str(os.path.basename(html_files_folder2[x]))+" from the public wifi --------------------------------------")
 			f.write("\r\r")
@@ -170,7 +179,7 @@ def _recursive_diff(folder1, folder2):
 	
 
 	
-	print('Job Completed, result.txt has been created')
+	print('Job Completed, result.txt, js_result.txt has been created')
 	return
 
 
@@ -180,9 +189,10 @@ def _recursive_diff_detail(folder1, folder2):
 	html_files_folder2= get_html_files(folder2)
 
 
-	f = open('result.txt','w')
+	f = open('detailed_result.txt','w')
 	if len(html_files_folder1) == len(html_files_folder2):
 		for x in range(0,len(html_files_folder1)-1):
+			print("Comparing "+ str(os.path.basename(html_files_folder1[x])) +"(file #1) and "+ str(os.path.basename(html_files_folder2[x]))+"(file #2) to see if any injected things inside the file #2.")
 			f.write("\r\r\r\r")
 			f.write("---[+][+][+]----Injected sources into-------"+str(os.path.basename(html_files_folder2[x]))+" from the public wifi --------------------------------------")
 			f.write("\r\r")
@@ -199,7 +209,7 @@ def _recursive_diff_detail(folder1, folder2):
 	
 
 	
-	print('Job Completed, result.txt has been created')
+	print('Job Completed, detailed_result.txt,js_result.txt has been created')
 	return
